refactor(meeting_backend/mod): extract tests to sibling (#1266)
Partial: mod.rs 988 -> 740 LOC (still over 400 — source split deferred).
new tests_mod.rs (249 LOC).
